Beta

+++ src/sck_server/sck_server_ops.c
@@ -351,6 +351,13 @@ void sck_server_op_open_ws ( sck_server_param_st * params, int sd, struct st_sck
   int  fd;
   char path [PATH_MAX];
 
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_open_ws] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_open_ws] >> Begin\n", params->rank);
 
   strcpy(path, params->dirbase);
@@ -370,6 +377,13 @@ void sck_server_op_open_wos ( sck_server_param_st * params, int sd, struct st_sc
 {
   int  fd;
   char path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_open_wos] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_open_wos] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_open_wos] open(%s)\n", params->rank, head->u_st_sck_server_msg.op_open.path);
@@ -394,6 +408,13 @@ void sck_server_op_creat_ws ( sck_server_param_st * params, int sd, struct st_sc
   int  fd;
   char path [PATH_MAX];
 
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_creat_ws] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_creat_ws] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_creat_ws] creat(%s)\n", params->rank, head->u_st_sck_server_msg.op_creat.path);
 
@@ -414,6 +435,13 @@ void sck_server_op_creat_wos ( sck_server_param_st * params, int sd, struct st_s
 {
   int  fd;
   char path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_creat_wos] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_creat_wos] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_creat_wos] creat(%s)\n", params->rank, head->u_st_sck_server_msg.op_creat.path);
@@ -438,6 +466,13 @@ void sck_server_op_read_ws ( sck_server_param_st * params, int sd, struct st_sck
   struct st_sck_server_read_req req;
   char * buffer;
   long size, diff, to_read, cont;
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_read_ws] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_read_ws] >> Begin\n",  params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_read_ws] read(%d, %ld %ld)\n", params->rank, head->u_st_sck_server_msg.op_read.fd, head->u_st_sck_server_msg.op_read.offset, head->u_st_sck_server_msg.op_read.size);
@@ -510,6 +545,13 @@ void sck_server_op_read_wos ( sck_server_param_st * params, int sd, struct st_sc
   char * buffer;
   long   size, diff, to_read, cont;
   char   path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_read_wos] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_read_wos] >> Begin\n",  params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_read_wos] read(%s, %ld %ld)\n", params->rank, head->u_st_sck_server_msg.op_read.path, head->u_st_sck_server_msg.op_read.offset, head->u_st_sck_server_msg.op_read.size);
@@ -601,6 +643,13 @@ void sck_server_op_write_ws ( sck_server_param_st * params, int sd, struct st_sc
   char * buffer;
   int    size, diff, cont, to_write;
 
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_write_ws] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_write_ws] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_write_ws] write(%d, %ld %ld)\n", params->rank, head->u_st_sck_server_msg.op_write.fd, head->u_st_sck_server_msg.op_write.offset, head->u_st_sck_server_msg.op_write.size);
 
@@ -661,6 +710,13 @@ void sck_server_op_write_wos ( sck_server_param_st * params, int sd, struct st_s
   char * buffer;
   int    size, diff, cont, to_write;
   char   path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_write_wos] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_write_wos] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_write_wos] write(%s, %ld %ld)\n", params->rank, head->u_st_sck_server_msg.op_write.path, head->u_st_sck_server_msg.op_write.offset, head->u_st_sck_server_msg.op_write.size);
@@ -736,14 +792,16 @@ void sck_server_op_write_wos ( sck_server_param_st * params, int sd, struct st_s
 
 void sck_server_op_close_ws ( sck_server_param_st * params, int sd, struct st_sck_server_msg * head, int rank_client_id )
 {
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_close_ws] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_close_ws] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_close_ws] close(%d)\n", params->rank, head->u_st_sck_server_msg.op_close.fd);
 
-  // check params...
-  if (NULL == params) {
-    return;
-  }
-  
   // do close
   int ret = filesystem_close(head->u_st_sck_server_msg.op_close.fd);
 
@@ -757,13 +815,15 @@ void sck_server_op_rm ( sck_server_param_st * params, int sd, struct st_sck_serv
 {
   char path [PATH_MAX];
 
-  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm] >> Begin\n", params->rank);
-  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm] unlink(%s)\n", params->rank, head->u_st_sck_server_msg.op_rm.path);
-
   // check params...
-  if (NULL == params) {
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_rm] ERROR: NULL arguments\n", -1);
     return;
   }
+
+  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm] >> Begin\n", params->rank);
+  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm] unlink(%s)\n", params->rank, head->u_st_sck_server_msg.op_rm.path);
 
   strcpy(path, params->dirbase);
   strcat(path, "/");
@@ -782,13 +842,15 @@ void sck_server_op_rm_async ( sck_server_param_st * params, __attribute__((__unu
 {
   char path [PATH_MAX];
 
-  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm_async] >> Begin\n", params->rank);
-  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm_async] unlink(%s)\n", params->rank, head->u_st_sck_server_msg.op_rm.path);
-
   // check params...
-  if (NULL == params) {
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_rm_async] ERROR: NULL arguments\n", -1);
     return;
   }
+
+  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm_async] >> Begin\n", params->rank);
+  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rm_async] unlink(%s)\n", params->rank, head->u_st_sck_server_msg.op_rm.path);
 
   strcpy(path, params->dirbase);
   strcat(path, "/");
@@ -806,13 +868,15 @@ void sck_server_op_rename ( sck_server_param_st * params, int sd, struct st_sck_
   char old_path [PATH_MAX];
   char new_path [PATH_MAX];
 
-  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rename] >> Begin\n", params->rank);
-  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rename] rename(%s, %s)\n", params->rank, head->u_st_sck_server_msg.op_rename.old_url, head->u_st_sck_server_msg.op_rename.new_url);
-
   // check params...
-  if (NULL == params) {
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_rename] ERROR: NULL arguments\n", -1);
     return;
   }
+
+  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rename] >> Begin\n", params->rank);
+  debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rename] rename(%s, %s)\n", params->rank, head->u_st_sck_server_msg.op_rename.old_url, head->u_st_sck_server_msg.op_rename.new_url);
 
   strcpy(old_path, params->dirbase);
   strcat(old_path, head->u_st_sck_server_msg.op_rename.old_url);
@@ -834,6 +898,13 @@ void sck_server_op_getattr ( sck_server_param_st * params, int sd, struct st_sck
   struct st_sck_server_attr_req req;
   char path [PATH_MAX];
 
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_getattr] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_getattr] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_getattr] stat(%s)\n", params->rank, head->u_st_sck_server_msg.op_getattr.path);
 
@@ -852,13 +923,16 @@ void sck_server_op_getattr ( sck_server_param_st * params, int sd, struct st_sck
 
 void sck_server_op_setattr ( sck_server_param_st * params, __attribute__((__unused__)) int sd, struct st_sck_server_msg * head, __attribute__((__unused__)) int rank_client_id )
 {
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_setattr] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_setattr] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_setattr] SETATTR(...)\n", params->rank);
 
-  // check params...
-  if (NULL == params) {
-    return;
-  }
   if (NULL == head) {
     return;
   }
@@ -875,6 +949,13 @@ void sck_server_op_mkdir ( sck_server_param_st * params, int sd, struct st_sck_s
 {
   int ret;
   char path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_mkdir] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_mkdir] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_mkdir] mkdir(%s)\n", params->rank, head->u_st_sck_server_msg.op_mkdir.path);
@@ -896,6 +977,13 @@ void sck_server_op_opendir ( sck_server_param_st * params, int sd, struct st_sck
 {
   DIR * ret;
   char path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_opendir] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_opendir] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_opendir] opendir(%s)\n", params->rank, head->u_st_sck_server_msg.op_opendir.path);
@@ -921,6 +1009,13 @@ void sck_server_op_readdir ( sck_server_param_st * params, int sd, struct st_sck
   struct dirent * ret;
   struct st_sck_server_direntry ret_entry;
   DIR * s;
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_readdir] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_readdir] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_readdir] readdir(%p)\n", params->rank, head->u_st_sck_server_msg.op_readdir.dir);
@@ -948,6 +1043,13 @@ void sck_server_op_closedir ( sck_server_param_st * params, int sd, struct st_sc
   int ret;
   DIR * s;
 
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_closedir] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_closedir] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_closedir] closedir(%p)\n", params->rank, head->u_st_sck_server_msg.op_closedir.dir);
 
@@ -965,6 +1067,13 @@ void sck_server_op_rmdir ( sck_server_param_st * params, int sd, struct st_sck_s
 {
   int ret;
   char path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_rmdir] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rmdir] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rmdir] rmdir(%s)\n", params->rank, head->u_st_sck_server_msg.op_rmdir.path);
@@ -986,6 +1095,13 @@ void sck_server_op_rmdir_async ( __attribute__((__unused__)) sck_server_param_st
 {
   int ret;
   char path [PATH_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_rmdir_async] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rmdir_async] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_rmdir_async] rmdir(%s)\n", params->rank, head->u_st_sck_server_msg.op_rmdir.path);
@@ -1017,6 +1133,13 @@ void sck_server_op_preload ( sck_server_param_st * params, int sd, struct st_sck
 
   int  BLOCKSIZE = head->u_st_sck_server_msg.op_preload.block_size;
   char buffer[BLOCKSIZE];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_preload] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_preload] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_preload] preload(%s,%s)\n", params->rank, head->u_st_sck_server_msg.op_preload.virtual_path, head->u_st_sck_server_msg.op_preload.storage_path);
@@ -1102,6 +1225,13 @@ void sck_server_op_flush ( sck_server_param_st * params, int sd, struct st_sck_s
   int BLOCKSIZE = head->u_st_sck_server_msg.op_flush.block_size;
   char buffer[BLOCKSIZE];
 
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_flush] ERROR: NULL arguments\n", -1);
+    return;
+  }
+
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_flush] >> Begin\n", params->rank);
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_flush] preload(%s,%s)\n", params->rank, head->u_st_sck_server_msg.op_flush.virtual_path, head->u_st_sck_server_msg.op_flush.storage_path);
 
@@ -1170,6 +1300,13 @@ void sck_server_op_flush ( sck_server_param_st * params, int sd, struct st_sck_s
 void sck_server_op_getnodename(sck_server_param_st * params, int sd, __attribute__((__unused__)) struct st_sck_server_msg * head, int rank_client_id)
 {
   char serv_name[HOST_NAME_MAX];
+
+  // check params...
+  if (NULL == params)
+  {
+    printf("[Server=%d] [MPI_SERVER_OPS] [sck_server_op_getnodename] ERROR: NULL arguments\n", -1);
+    return;
+  }
 
   debug_info("[Server=%d] [SCK_SERVER_OPS] [sck_server_op_getnodename] >> Begin\n", params->rank);
